create a class dedicated for checking whether a work order needs to be reapproved
